refactor: extract shared word-wrap algorithm to wrap_util (#527)
scroll_buffer::visual_height() and wrap_input::visual_line_count() were
identical implementations of the same word-boundary wrapping algorithm.

- Create wrap_util.rs as the single source of truth
- Both callers now delegate to wrap_util::visual_line_count()
- Add comprehensive tests in wrap_util covering all edge cases
- Existing tests in scroll_buffer and wrap_input continue to pass
  (they exercise the delegation path)

DRY violation identified by code review.
